Hydrate one PGlite per test file, and roll each test back
`seededTestDb()` hydrates a fresh embedded Postgres per test at 0.6-0.9s each,
and that is most of the unit suite's wall-clock: the 38 database-backed files
under `src/db` carried 849 of 988 test-seconds in a measured CI shard. The
template already made this ten times cheaper than migrating per test; what is
left is structural.

`fileScopedShopContext()` hydrates once per file and gives each test a
transaction that is rolled back afterwards — the same isolation arrived at
differently, typed as `AppDb` so query code needs no change. Drizzle's
`transaction` is callback-scoped, so the transaction is held open by parking
its callback on a promise the teardown resolves, then thrown out with a
sentinel to roll back.

Measured across the three converted files (132 tests):

    before   109.28s test time,  48.77s wall-clock
    after     12.78s test time,  11.23s wall-clock

`trips-queries.test.ts` alone, which the issue asks for by name: 17.33s to
1.71s.

Which files can take it was decided per file, not assumed, and two declined for
reasons worth writing down.

`manifests.test.ts` stays on per-test hydration: two of its tests have
transaction semantics as their *subject* — they open their own transaction so a
pair of roll-call rows share one `now()`, then assert the append-order
tie-break. Under a file-scoped transaction the inner one becomes a savepoint
and `now()` is already frozen, so every row in the test would carry the same
timestamp and the pair would stop being the thing under test. The assertion
would still pass, and prove less.

`divers.test.ts`'s erasure test takes its own database while the file's other
29 tests share one, and this is the case I did not see coming: it provokes a
database error on purpose (`.rejects.toThrow()` on a restore a constraint
refuses) and then keeps querying. Postgres aborts the *whole* transaction on a
failed statement — 25P02, "commands ignored until end of transaction block" —
so inside a transaction every line after the expected rejection fails too. In
autocommit only the failing statement fails, which is what the assertion was
written against. It failed exactly that way on the first run.

Both, plus `FOR UPDATE` races and the money paths, are written up in
`fileScopedShopContext`'s own doc comment and in
`docs/engineering/testing.md`, with the rule short enough to apply: read-heavy
files that commit nothing. A file that declines says so where it declines,
naming which case applies.

`pnpm test src/db src/test` is green — 2,079 passed, 22 skipped.

Closes #1244

+++ src/test/db-template.test.ts
@@ -56,7 +56,7 @@ describe("templateBytes", () => {
     const templateBytes = await freshTemplateBytes();
     const bytes = await templateBytes("lean");
     expect(bytes).not.toBeNull();
-    expect(Buffer.from(bytes).toString()).toBe("not-really-a-tar");
+    expect(Buffer.from(bytes as Uint8Array).toString()).toBe("not-really-a-tar");
   });
 
   it("answers null when there is no snapshot, so the caller can seed from scratch", async () => {
@@ -82,7 +82,7 @@ describe("templateBytes", () => {
     await writeFile(path.join(cacheDir, "test-db-template-history.tar"), Buffer.from("history"));
     const templateBytes = await freshTemplateBytes();
     expect(await templateBytes("lean")).toBeNull();
-    expect(Buffer.from(await templateBytes("history")).toString()).toBe("history");
+    expect(Buffer.from((await templateBytes("history")) as Uint8Array).toString()).toBe("history");
   });
 });
